[ObjC]fix 64 to 32 bit clang conversion warning in src/core/lib (#29877)

+++ src/core/lib/surface/channel.h
@@ -145,7 +145,7 @@ class Channel : public RefCounted<Channel>,
   bool is_promising() const { return is_promising_; }
   RegisteredCall* RegisterCall(const char* method, const char* host);
 
-  int TestOnlyRegisteredCalls() {
+  size_t TestOnlyRegisteredCalls() {
     MutexLock lock(&registration_table_.mu);
     return registration_table_.map.size();
   }
